requirements.txt
vosk
opensmile
numpy
sentence-transformers
sounddevice
django
djangorestframework
gunicorn
